refactor: move provision steps into separate file
Moved the steps from `ProvisionSteps` into a separate file.
